% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarcoSim-package.R
\docType{package}
\name{BarcoSim-package}
\alias{BarcoSim}
\alias{BarcoSim-package}
\title{BarcoSim: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
