% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_similarity.R
\name{calSeqSim}
\alias{calSeqSim}
\title{Calculate sequence similarity}
\usage{
calSeqSim(sequences)
}
\arguments{
\item{sequences}{A character vector of sequences to compare.}
}
\value{
A numeric vector of similarity values for each position.
}
\description{
This function calculates the similarity between sequences at each position.
}
\examples{
# A dataset of DNA sequences

sequences <- c("CGCCTCCACG", "CGAGCACACG", "CGGAGTCACG", "CGCCAACACG", "CGTAAGCACG",
               "CGCCAACACG", "CGTGCGCACG", "CGCTGTCACG", "CGGCCACACG", "CGGCGTCACG")

# Checking at the similarity among them

similarity <- calSeqSim(sequences)
similarity

}
