% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_gpseq_custom.R
\name{r_gpseq_csub}
\alias{r_gpseq_csub}
\title{Create barcode sequences from parent sequences but with non random substitutions}
\usage{
r_gpseq_csub(dna_seq, num_replicates, substitution_probs)
}
\arguments{
\item{dna_seq}{A character, output from gpseq}

\item{num_replicates}{A numeric value}

\item{substitution_probs}{A list of probabilities for each based pair or an indel to appear}
}
\value{
A data.frame
}
\description{
#for each the parent sequence coming from the gpseq function, create n offspring sequences with a substitution rate of your interest.
}
\examples{
dna_seq <- c("AAGA","AATC")
substitution_probs <- list("A" = 0.1, "C" = 0.2, "G" = 0.3, "T" = 0.4, " " = 0.1)
r_gpseq_csub(dna_seq,3,substitution_probs)

}
