% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parent_sequences.R
\name{gpseq}
\alias{gpseq}
\title{Generate parent sequences}
\usage{
gpseq(num_sequences, seq_length, range_start, range_end)
}
\arguments{
\item{num_sequences}{A numeric value}

\item{seq_length}{A numeric value}

\item{range_start}{A numeric value}

\item{range_end}{A numeric value}
}
\value{
A character vector, DNA sequences.
}
\description{
`gpseq` creates DNA sequences of certain length. These sequences are indetical apart from a specific range that refers to barcodes and its variable.
}
\examples{
# Create 4 DNA sequneces, 150 bp each, variable in the are between 53-78.
gpseq(4,150,53,78)

}
