% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_gpseq_custom.R
\name{r_gpseq_csub}
\alias{r_gpseq_csub}
\title{Create barcode sequences from parent sequences with non-random substitutions.}
\usage{
r_gpseq_csub(
  dna_seq,
  num_replicates,
  error_rate,
  substitution_probs,
  color = "no"
)
}
\arguments{
\item{dna_seq}{A character vector of parent DNA sequences.}

\item{num_replicates}{An integer specifying the number of offspring sequences per parent.}

\item{error_rate}{A numeric value (0-1) indicating the likelihood of a substitution at each base.}

\item{substitution_probs}{A named list indicating the substitution probabilities for each base (A, C, G, T, and '').}

\item{color}{A character flag. If set to "yes", the function will output color-coded sequences.}
}
\value{
A data frame containing the parent sequence, the offspring sequences, and parent index.
        If color is "yes", sequences are printed with color-coding and the function returns NULL.
}
\description{
Given parent DNA sequences, this function generates offspring sequences with potential substitutions.
Substitutions are determined by an error rate and a specified set of substitution probabilities.
}
\examples{
dna_seq <- c("AAGA", "AATC")
num_replicates <- 3
error_rate <- 0.1
substitution_probs <- list("A" = 0.1, "C" = 0.2, "G" = 0.3, "T" = 0.3, " " = 0.1)
barcode_sequences <- r_gpseq_csub(dna_seq,
 num_replicates, error_rate, substitution_probs,
 color = "yes")

}
